pyproject.toml
[tool.poetry]
name = "hw24-1-project"
version = "0.1.0"
description = ""
authors = ["rinatyuk-git <17_ufo_RB@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django-admin = "^2.0.2"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
